requirements.txt
Django==1.8.6
django-authtools==1.4.0
django-twitter-bootstrap==3.3.0
